api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.4
openai==0.28.1
